reincarnation release
`with-editor.el' is no longer part of the Magit repository but its
karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit https://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> f78f1cd33692e8b544d883d66400ab5c752663f2

See the git-replace(1) manpage for more information.

While they were still part of the Magit repository `with-editor.el'
and `with-editor.org' were edited by the following people:

    88  Jonas Bernoulli
     4  Noam Postavsky
     2  Rémi Vanicat
     1  Barak A. Pearlmutter
     1  Lele Gaifax
     1  Philippe Vaucher