with-editor-finish: Fix bug concerning git-commit-post-finish-hook
`git-commit-post-finish-hook' is intended to be set locally.  Because
the buffer in which it is locally does not exist anymore by the time
we have reached the "post finish" time, we use a temporary buffer and
carry over the value from the other buffer.

However we accidentally only copied the value of the hook variable if
it already non-nil locally in the temporary buffer (which should not
be the case) or non-nil globally (which might very well not be the
case.

Closes #68.