with-editor-export-editor: Fix for term-mode
The previous value of `with-editor-sleeping-editor' variable was using
the \x1f character in the elisp string and sending it to the
shell/term process.

For `shell-mode' this is not an issue since the character is not
interpreted by the shell. But in `term-mode' the terminal emulation
interprets it and remove previous characters, the "$0" string. So we
end up with an EDITOR environment variable that contains an echo
command with "…OPEN IN…" instead of "…OPEN $0^_ IN…" (here "^_" is
used to represent the \x1f byte).

This change fixes the issue by using `-e' flag of `echo' and using the
`\037' octal escape sequence for the \x1f byte.