with-editor-export-editor: Use printf instead of echo
The `-e' flag of `echo' that makes the command interpret
backslash-escaped sequences (in particular "\037") is not supported
in all shells.

Shells that do not support it print the flag instead.

So here we replace `echo' by `printf' which seems to be better
supported and does the interpretation of backslash-escaped sequence
by default.

Re #71.
Closes #72.