with-editor-sleeping-editor-filter: Limit length of saved output
Improve performance by limitting the maximum length of the string
saved in process filters.

While using with `vterm', this `incomplete' string may get very
large and it hurts performance.  In my case, it can easily get to few
megabytes after using the same terminal for a long time, which would
then increase the latency of each keystroke significantly.  We should
only need to keep the last few bytes to make it work.