with-editor-mode: Prevent dangerous use as a command
We already prevent disabling the mode interactively, but now we do
so by not even defining it as a command, if the used Emacs version
supports it.

More importantly, also prevent the mode from being enabled.  Enabling
the mode manually is dangerous because, as always, that instructs the
user on how to cancel the buffer, and if they do that, then that also
deletes the visited file.

Closes #131.